chore: splitting payment method hooks
