+++ src/components/Employee/Dashboard/hooks/useEmployeeCompensation.tsx
@@ -1,23 +1,15 @@
 import { useMemo } from 'react'
 import { useEmployeesGetSuspense } from '@gusto/embedded-api/react-query/employeesGet'
-import { useEmployeePaymentMethodGetSuspense } from '@gusto/embedded-api/react-query/employeePaymentMethodGet'
-import { useEmployeePaymentMethodsGetBankAccountsSuspense } from '@gusto/embedded-api/react-query/employeePaymentMethodsGetBankAccounts'
 import { useGarnishmentsListSuspense } from '@gusto/embedded-api/react-query/garnishmentsList'
 import { usePayrollsGetPayStubsSuspense } from '@gusto/embedded-api/react-query/payrollsGetPayStubs'
 import type { Job } from '@gusto/embedded-api/models/components/job'
 import type { Garnishment } from '@gusto/embedded-api/models/components/garnishment'
-import type { GetV1EmployeesEmployeeIdPaymentMethodResponse } from '@gusto/embedded-api/models/operations/getv1employeesemployeeidpaymentmethod'
-import type { GetV1EmployeesEmployeeIdBankAccountsResponse } from '@gusto/embedded-api/models/operations/getv1employeesemployeeidbankaccounts'
 import type { GetV1EmployeesEmployeeUuidPayStubsResponse } from '@gusto/embedded-api/models/operations/getv1employeesemployeeuuidpaystubs'
 import { composeErrorHandler } from '@/partner-hook-utils/composeErrorHandler'
 import { usePagination } from '@/hooks/usePagination/usePagination'
 import type { HookLoadingResult, BaseHookReady } from '@/partner-hook-utils/types'
 import type { PaginationControlProps } from '@/components/Common/PaginationControl/PaginationControlTypes'
 
-// Derive types from operations responses
-type EmployeeBankAccount = NonNullable<
-  GetV1EmployeesEmployeeIdBankAccountsResponse['employeeBankAccounts']
->[number]
 type EmployeePayStub = NonNullable<
   GetV1EmployeesEmployeeUuidPayStubsResponse['employeePayStubsList']
 >[number]
@@ -29,10 +21,6 @@ export interface UseEmployeeCompensationProps {
 interface UseEmployeeCompensationReady extends BaseHookReady<
   {
     primaryJob?: Job
-    employeePaymentMethod?: NonNullable<
-      GetV1EmployeesEmployeeIdPaymentMethodResponse['employeePaymentMethod']
-    >
-    bankAccounts: EmployeeBankAccount[]
     garnishmentList: Garnishment[]
     payStubs: EmployeePayStub[]
   },
@@ -53,8 +41,6 @@ export function useEmployeeCompensation({
   })
 
   const employeeQuery = useEmployeesGetSuspense({ employeeId })
-  const paymentMethodQuery = useEmployeePaymentMethodGetSuspense({ employeeId })
-  const bankAccountsQuery = useEmployeePaymentMethodsGetBankAccountsSuspense({ employeeId })
   const garnishmentsQuery = useGarnishmentsListSuspense({ employeeId })
   const payStubsQuery = usePayrollsGetPayStubsSuspense({
     employeeId,
@@ -63,25 +49,15 @@ export function useEmployeeCompensation({
   })
 
   const employee = employeeQuery.data.employee
-  const employeePaymentMethod = paymentMethodQuery.data.employeePaymentMethod
-  const bankAccountsData = bankAccountsQuery.data
   const garnishmentList = garnishmentsQuery.data.garnishments
   const payStubsData = payStubsQuery.data
 
-  // Derive primary job
   const primaryJob = useMemo(() => {
     return employee?.jobs?.find(job => job.primary === true)
   }, [employee?.jobs])
 
-  // Derive bank accounts
-  const bankAccounts = useMemo(() => {
-    return bankAccountsData.employeeBankAccounts || []
-  }, [bankAccountsData])
-
-  // Extract paystubs from response
   const payStubs = payStubsData.employeePayStubsList || []
 
-  // Extract pagination from API response headers
   const payStubsPagination = useMemo(() => {
     const headers = payStubsData.httpMeta.response.headers
     return getPaginationProps(headers, payStubsQuery.isFetching)
@@ -89,8 +65,6 @@ export function useEmployeeCompensation({
 
   const isPending =
     employeeQuery.isFetching ||
-    paymentMethodQuery.isFetching ||
-    bankAccountsQuery.isFetching ||
     garnishmentsQuery.isFetching ||
     payStubsQuery.isFetching
 
@@ -98,8 +72,6 @@ export function useEmployeeCompensation({
 
   const errorHandling = composeErrorHandler([
     employeeQuery,
-    paymentMethodQuery,
-    bankAccountsQuery,
     garnishmentsQuery,
     payStubsQuery,
   ])
@@ -115,8 +87,6 @@ export function useEmployeeCompensation({
     isLoading: false,
     data: {
       primaryJob,
-      employeePaymentMethod: employeePaymentMethod || undefined,
-      bankAccounts,
       garnishmentList: garnishmentList || [],
       payStubs,
     },
