+++ src/components/Employee/PaymentMethod/management/index.ts
@@ -1,0 +1,3 @@
+export { ListView } from './ListView'
+export { PaymentMethod } from './PaymentMethod'
+export type { PaymentMethodProps } from './PaymentMethod'
