+++ src/components/Employee/PaymentMethod/shared/index.ts
@@ -1,3 +1,10 @@
-export * from './PaymentMethodComponents'
-export * from './usePaymentMethod'
 export * from './paymentMethodSchema'
+export * from './BankAccountFormFields'
+export * from './SplitViewFields'
+export * from './useSplitViewState'
+export * from './useDeleteBankAccount'
+export * from './DeleteBankAccountDialog'
+export * from './usePaymentMethodList'
+export * from './usePaymentMethodForm'
+export * from './useBankForm'
+export * from './useSplitPaymentsForm'
